+++ pyinstaller.txt
@@ -1,5 +1,5 @@
-pyinstaller -F ./Sakura_Launcher_GUI.py --icon=icon.png --add-data "icon.png;." -w --hidden-import=tiktoken_ext.openai_public --hidden-import=tiktoken_ext --hidden-import=PySide6 --hidden-import=PySide6.QtCore --hidden-import=PySide6.QtGui --hidden-import=qfluentwidgets --clean
+pyinstaller -F ./src/main.py --icon=icon.png --add-data "icon.png;." -w --hidden-import=tiktoken_ext.openai_public --hidden-import=tiktoken_ext --hidden-import=PySide6 --hidden-import=PySide6.QtCore --hidden-import=PySide6.QtGui --hidden-import=qfluentwidgets --clean
 
-pyinstaller -F ./Sakura_Launcher_GUI.py --icon=icon.png --add-data "icon.png;." --hidden-import=tiktoken_ext.openai_public --hidden-import=tiktoken_ext --hidden-import=PySide6 --hidden-import=PySide6.QtCore --hidden-import=PySide6.QtGui --hidden-import=qfluentwidgets --clean
+pyinstaller -F ./src/main.py --icon=icon.png --add-data "icon.png;." --hidden-import=tiktoken_ext.openai_public --hidden-import=tiktoken_ext --hidden-import=PySide6 --hidden-import=PySide6.QtCore --hidden-import=PySide6.QtGui --hidden-import=qfluentwidgets --clean
 
-pyinstaller -F ./Sakura_Launcher_GUI.py --icon=icon.png --add-data "icon.png;." --add-data "cloudflared-windows-amd64.exe;." -w --hidden-import=tiktoken_ext.openai_public --hidden-import=tiktoken_ext --hidden-import=PySide6 --hidden-import=PySide6.QtCore --hidden-import=PySide6.QtGui --hidden-import=qfluentwidgets --clean
+pyinstaller -F ./src/main.py --icon=icon.png --add-data "icon.png:." --add-data "cloudflared-windows-amd64.exe:." -w --hidden-import=tiktoken_ext.openai_public --hidden-import=tiktoken_ext --hidden-import=PySide6 --hidden-import=PySide6.QtCore --hidden-import=PySide6.QtGui --hidden-import=qfluentwidgets --clean
